+++ src/components/common/vatsim/CommonSingleControllerInfo.vue
@@ -100,7 +100,13 @@
             #bottom
         >
             <div>
-                Booked from {{ controller.booking?.start_local }} to {{ controller.booking?.end_local }}
+                Booked from
+                <template v-if="store.mapSettings.bookingsLocalTimezone">
+                    {{ controller.booking?.start_local }} to {{ controller.booking?.end_local }}
+                </template>
+                <template v-else>
+                    {{ controller.booking?.start_z }}Z to {{ controller.booking?.end_z }}Z
+                </template>
             </div>
         </template>
     </common-info-block>
@@ -117,6 +123,7 @@ import CommonAtcTimeOnline from '~/components/common/vatsim/CommonAtcTimeOnline.
 import CommonSpoiler from '~/components/common/vatsim/CommonSpoiler.vue';
 import SaveIcon from '@/assets/icons/kit/save.svg?component';
 import { getStringColorFromSettings } from '~/composables/colors';
+import { useStore } from '~/store';
 
 defineProps({
     controller: {
@@ -159,6 +166,7 @@ const dataStore = useDataStore();
 const mapStore = useMapStore();
 const { copy, copyState } = useCopyText();
 const copiedFor = ref('');
+const store = useStore();
 
 const controllerColor = (controller: VatsimShortenedController) => {
     const list = getUserList(controller.cid);